Day3_Basics_of_GenAI/Task2_Ocean_Poem_HF/requirements.txt
streamlit
transformers
groq
python-dotenv
